Filtrado con imagenes float32

+++ comandos5.py
@@ -63,13 +63,12 @@ def dotplot_chunk(chunk, Secuencia1, Secuencia2):
 
 def filter_dotplot(dotplot):
     window_size = 5
-    window = np.eye(window_size, dtype=np.float16)
-    filtered_dotplot = convolve2d(
-        dotplot, window, mode='same', dtype=np.float16)
+    window = np.eye(window_size, dtype=np.float32)
+    filtered_dotplot = convolve2d(dotplot, window, mode='same')
+    filtered_dotplot = filtered_dotplot.astype(np.float16)
     threshold = np.float16(0.8 * window_size)
     binary_dotplot = (filtered_dotplot >= threshold).astype(np.int8)
     return binary_dotplot
-
 # Función para calcular el dotplot utilizando mpi4py
